created Holiday controller and added changes to DTO and service

+++ GroceryStore-Backend/src/main/java/ca/mcgill/ecse321/GroceryStore/service/HolidayService.java
@@ -40,7 +40,11 @@ public class HolidayService {
     }
     @Transactional
     public void deleteHoliday(String name){
-        holidayRepository.deleteById(name);
+        if (name == null || name.equals("")) {
+            throw new IllegalArgumentException("The business hour with ID: " + name + " does not exist.");
+        } else {
+            holidayRepository.deleteById(name);
+        }
     }
